+++ tests/useInnerWatch/watching-multiple-stores.spec.ts
@@ -2,7 +2,7 @@ import { useCallback } from "react"
 import { act, renderHook } from "@testing-library/react-hooks"
 
 import { Compare, InnerStore, useInnerWatch, batch } from "../../src"
-import { Counter } from "../helpers"
+import { Counter } from "../common"
 
 interface WithStore<T = Counter> {
   store: InnerStore<T>
