+++ tests/useSetInnerState.spec.ts
@@ -3,7 +3,7 @@ import { act, renderHook } from "@testing-library/react-hooks"
 
 import { InnerStore, SetInnerState, useSetInnerState } from "../src"
 
-import { Counter } from "./helpers"
+import { Counter } from "./common"
 
 describe("bypassed store", () => {
   it.concurrent.each([
